+++ src/apkcontroller/helper.py
@@ -2,33 +2,18 @@
 import json
 import os
 import subprocess  # nosec
+import time
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, List, Union
+from typing import TYPE_CHECKING, Dict, List, Optional, Union
 
+import xmltodict
 from typeguard import typechecked
 from uiautomator import AutomatorDevice
-
-from src.apkcontroller.script_helper import (
-    get_screen_as_dict,
-    is_expected_screen,
-)
 
 if TYPE_CHECKING:
     from src.apkcontroller.Screen import Screen
 else:
     Screen = object
-
-
-@typechecked
-def file_exists(filepath: str) -> bool:
-    """Checks if file exists.
-
-    :param string:
-    """
-    # TODO: Execute Path(string).is_file() directly instead of calling this
-    # function.
-    my_file = Path(filepath)
-    return my_file.is_file()
 
 
 @typechecked
@@ -102,7 +87,7 @@ def export_screen_data_if_valid(
                     screen_dict=screen.screen_dict,
                     script_description=screen.script_description,
                     overwrite=overwrite,
-                    unverified=False,
+                    subdir="verified",
                 )
 
 
@@ -141,8 +126,6 @@ def run_bash_command(
 def launch_app(app_name: str) -> None:
     """Launches app on phone."""
 
-    # TODO: verify phone is connected.
-
     # Launc the app on phone.
     command = f'adb shell monkey -p "{app_name}" 1 &>/dev/null'
     run_bash_command(
@@ -158,26 +141,22 @@ def export_screen_data(
     screen_dict: Dict,
     script_description: Dict,
     overwrite: bool = False,
-    unverified: bool = True,
+    subdir: str = "unverified",
 ) -> None:
     """Writes a dict file to a .json file, and exports a screenshot.
 
     The overwrite bool determines whether a file will be overwritten or
-    not if it  already exists. The unverified bool determines whether
-    the screen data is verified to be expected, or not. If not, the
-    screen data is placed in a subfolder named: unverified, to reduce
-    the probability of the developer basing script actions on data
-    belonging to the wrong screen.
+    not if it  already exists. The subdir bool determines whether the
+    screen data is verified to be expected, or not. If not, the screen
+    data is placed in a subfolder named: subdir, to reduce the
+    probability of the developer basing script actions on data belonging
+    to the wrong screen.
     """
-    if unverified:
-        unverified_dir = "unverified/"
-    else:
-        unverified_dir = "verified/"
     output_dir = (
         (
             "src/apkcontroller/"
             + f'{script_description["app_name"]}'
-            + f'/V{script_description["version"]}/{unverified_dir}'
+            + f'/V{script_description["version"]}/{subdir}'
         )
         .replace(".", "_")
         .replace(" ", "_")
@@ -204,3 +183,145 @@ def export_screen_data(
         # Verify the file exists.
         if not Path(output_path).is_file():
             raise Exception(f"Error, filepath:{output_path} was not created.")
+
+
+@typechecked
+def get_screen_as_dict(
+    device: AutomatorDevice,
+    unpack: bool,
+    screen_dict: Dict,
+    reload: bool = False,
+) -> Dict:
+    """Loads the phone and shows the screen as a dict."""
+
+    # Don't reload if the screen dict still exists, and no explicit
+    # reload is asked.
+    if screen_dict != {} and not reload:
+        if unpack and "hierarchy" in screen_dict.keys():
+            return screen_dict["hierarchy"]
+        return screen_dict
+
+    # Get the new screen data from the ui.
+    if screen_dict == {} or reload:
+        print("Loading screen data from phone for orientation.")
+        ui_information: Dict = xmltodict.parse(device.dump())
+
+        # Unpack the screen dict to get a recursive dictionary structure.
+        if unpack:
+            ui_information = ui_information["hierarchy"]
+    return ui_information
+
+
+@typechecked
+def is_expected_screen(
+    device: AutomatorDevice,
+    expected_screen: Screen,
+    retry: bool,
+    unpacked_screen_dict: Dict,
+    verbose: Optional[bool] = False,
+) -> bool:
+    """Custom verification per screen based on the optional and required
+    objects in screen.
+
+    Raise error if verification fails.
+    """
+    # Preliminary check to see if the required objects are in.
+    if not required_objects_in_screen(
+        expected_screen.required_objects, unpacked_screen_dict
+    ):
+        if not retry:
+            return False
+        # Retry and return True if the required objects were found.
+        for _ in range(0, expected_screen.max_retries):
+
+            # Reload the screen data again.
+            unpacked_screen_dict = get_screen_as_dict(
+                device=device,
+                unpack=True,
+                screen_dict={},
+                reload=True,
+            )
+
+            time.sleep(expected_screen.wait_time_sec)
+            if required_objects_in_screen(
+                required_objects=expected_screen.required_objects,
+                unpacked_screen_dict=unpacked_screen_dict,
+            ):
+                return True
+            if verbose:
+                print(f"Not found:{expected_screen.required_objects}")
+        if verbose:
+            print(f"Not found:{expected_screen.required_objects}")
+        return False
+    # else:
+    return True
+
+
+@typechecked
+def required_objects_in_screen(
+    required_objects: List[Dict[str, Union[List, Dict, str]]],
+    unpacked_screen_dict: Dict[str, Union[List, Dict, str]],
+) -> bool:
+    """Returns True if all required objects are found in the UI/screen.
+
+    False otherwise.
+    """
+    for required_object in required_objects:
+        if not required_object_in_screen(
+            required_object=required_object,
+            unpacked_screen_dict=unpacked_screen_dict,
+        ):
+            return False
+    return True
+
+
+@typechecked
+def required_object_in_screen(
+    required_object: Dict[str, Union[List, Dict, str]],
+    unpacked_screen_dict: Dict[str, Union[List, Dict, str]],
+) -> bool:
+    """Returns True if all the keys and values in a dict are found within the
+    same key or list of the xml of the ui."""
+    if dict_contains_other_dict(required_object, unpacked_screen_dict):
+        return True
+    if "node" in unpacked_screen_dict.keys():
+        if isinstance(unpacked_screen_dict["node"], Dict):
+            if required_object_in_screen(
+                required_object=required_object,
+                unpacked_screen_dict=unpacked_screen_dict["node"],
+            ):
+                return True
+        if isinstance(unpacked_screen_dict["node"], List):
+            for node_elem in unpacked_screen_dict["node"]:
+                if required_object_in_screen(
+                    required_object=required_object,
+                    unpacked_screen_dict=node_elem,
+                ):
+                    return True
+        if not isinstance(unpacked_screen_dict["node"], Dict | List):
+            raise TypeError("Node value of unexpected type.")
+    return False
+
+
+@typechecked
+def dict_contains_other_dict(sub: Dict, main: Dict) -> bool:
+    """Returns true if the sub dict is a subset of the main dict."""
+    for sub_key, sub_val in sub.items():
+        if sub_key not in main.keys():
+            return False
+        # An artifact like:
+        # "@text": "VPN Mode \u200e\u200f\u200e\u200e\u200e\u200e\u200e\u200f
+        # \u200e\u200f\u200f\u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\
+        # u200e\u200e\u200f\u200e\u200e\u200e\u200e\u200f\u200f\u200f\u200f\
+        # u200f\u200e\u200e\u200f\u200f\u200e\u200e\u200e\u200f\u200e\u200e\
+        # u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200f\u200f\
+        # u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200f\u200e\
+        # u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200e\u200e\u200e\
+        # u200f\u200e\u200e\u200e\u200f\u200f\u200f\u200f\u200f\u200e\u200e\
+        # u200f\u200f\u200e\u200f\u200f\u200e\u200e\u200e\u200eON\u200e\u200f
+        # \u200e\u200e\u200f\u200e",
+        # may occur at random. Therefore, the "in" option is included.
+        # TODO: determine why these artifacts may occur and/or remove them.
+        if sub_val != main[sub_key] and sub_val not in main[sub_key]:
+            return False
+    return True
