+++ src/apkcontroller/script_helper.py
@@ -1,12 +1,16 @@
 """Functions to assist a script file for an arbitrary app."""
-import time
-from pprint import pprint
-from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, Union, cast
+from typing import TYPE_CHECKING, Dict, List, Tuple, Union, cast
 
 import networkx as nx
-import xmltodict
 from typeguard import typechecked
 from uiautomator import AutomatorDevice
+
+from src.apkcontroller.helper import (
+    dict_contains_other_dict,
+    export_screen_data,
+    get_screen_as_dict,
+    is_expected_screen,
+)
 
 if TYPE_CHECKING:
     from src.apkcontroller.org_torproject_android.V16_6_3_RC_1.script import (
@@ -67,124 +71,6 @@ def current_screen_is_expected(
                 int(str(expected_screen.script_description["screen_nr"])),
             )
     return (False, -1)
-
-
-@typechecked
-def is_expected_screen(
-    device: AutomatorDevice,
-    expected_screen: Screen,
-    retry: bool,
-    unpacked_screen_dict: Dict,
-    verbose: Optional[bool] = False,
-) -> bool:
-    """Custom verification per screen based on the optional and required
-    objects in screen.
-
-    Raise error if verification fails.
-    """
-    # Preliminary check to see if the required objects are in.
-    if not required_objects_in_screen(
-        expected_screen.required_objects, unpacked_screen_dict
-    ):
-        if not retry:
-            return False
-        # Retry and return True if the required objects were found.
-        for _ in range(0, expected_screen.max_retries):
-
-            # Reload the screen data again.
-            unpacked_screen_dict = get_screen_as_dict(
-                device=device,
-                unpack=True,
-                screen_dict={},
-                reload=True,
-            )
-
-            time.sleep(expected_screen.wait_time_sec)
-            if required_objects_in_screen(
-                required_objects=expected_screen.required_objects,
-                unpacked_screen_dict=unpacked_screen_dict,
-            ):
-                return True
-            if verbose:
-                print(f"Not found:{expected_screen.required_objects}")
-        if verbose:
-            print(f"Not found:{expected_screen.required_objects}")
-        return False
-    # else:
-    return True
-
-
-@typechecked
-def required_objects_in_screen(
-    required_objects: List[Dict[str, Union[List, Dict, str]]],
-    unpacked_screen_dict: Dict[str, Union[List, Dict, str]],
-) -> bool:
-    """Returns True if all required objects are found in the UI/screen.
-
-    False otherwise.
-    """
-    for required_object in required_objects:
-        if not required_object_in_screen(
-            required_object=required_object,
-            unpacked_screen_dict=unpacked_screen_dict,
-        ):
-            return False
-    return True
-
-
-@typechecked
-def get_screen_as_dict(
-    device: AutomatorDevice,
-    unpack: bool,
-    screen_dict: Dict,
-    reload: bool = False,
-) -> Dict:
-    """Loads the phone and shows the screen as a dict."""
-
-    # Don't reload if the screen dict still exists, and no explicit
-    # reload is asked.
-    if screen_dict != {} and not reload:
-        if unpack and "hierarchy" in screen_dict.keys():
-            return screen_dict["hierarchy"]
-        return screen_dict
-
-    # Get the new screen data from the ui.
-    if screen_dict == {} or reload:
-        print("Loading screen data from phone for orientation.")
-        ui_information: Dict = xmltodict.parse(device.dump())
-
-        # Unpack the screen dict to get a recursive dictionary structure.
-        if unpack:
-            ui_information = ui_information["hierarchy"]
-    return ui_information
-
-
-@typechecked
-def required_object_in_screen(
-    required_object: Dict[str, Union[List, Dict, str]],
-    unpacked_screen_dict: Dict[str, Union[List, Dict, str]],
-) -> bool:
-    """Returns True if all the keys and values in a dict are found within the
-    same key or list of the xml of the ui."""
-    if dict_contains_other_dict(required_object, unpacked_screen_dict):
-        return True
-    if "node" in unpacked_screen_dict.keys():
-        if isinstance(unpacked_screen_dict["node"], Dict):
-            if required_object_in_screen(
-                required_object=required_object,
-                unpacked_screen_dict=unpacked_screen_dict["node"],
-            ):
-                return True
-        if isinstance(unpacked_screen_dict["node"], List):
-            for node_elem in unpacked_screen_dict["node"]:
-                if required_object_in_screen(
-                    required_object=required_object,
-                    unpacked_screen_dict=node_elem,
-                ):
-                    return True
-        if not isinstance(unpacked_screen_dict["node"], Dict | List):
-            raise TypeError("Node value of unexpected type.")
-    return False
 
 
 @typechecked
@@ -252,30 +138,6 @@ def get_torified_item_index_dict(
 
 
 @typechecked
-def dict_contains_other_dict(sub: Dict, main: Dict) -> bool:
-    """Returns true if the sub dict is a subset of the main dict."""
-    for sub_key, sub_val in sub.items():
-        if sub_key not in main.keys():
-            return False
-        # An artifact like:
-        # "@text": "VPN Mode \u200e\u200f\u200e\u200e\u200e\u200e\u200e\u200f
-        # \u200e\u200f\u200f\u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\
-        # u200e\u200e\u200f\u200e\u200e\u200e\u200e\u200f\u200f\u200f\u200f\
-        # u200f\u200e\u200e\u200f\u200f\u200e\u200e\u200e\u200f\u200e\u200e\
-        # u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200f\u200f\
-        # u200f\u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200f\u200e\
-        # u200e\u200e\u200e\u200e\u200e\u200e\u200f\u200e\u200e\u200e\u200e\
-        # u200f\u200e\u200e\u200e\u200f\u200f\u200f\u200f\u200f\u200e\u200e\
-        # u200f\u200f\u200e\u200f\u200f\u200e\u200e\u200e\u200eON\u200e\u200f
-        # \u200e\u200e\u200f\u200e",
-        # may occur at random. Therefore, the "in" option is included.
-        # TODO: determine why these artifacts may occur and/or remove them.
-        if sub_val != main[sub_key] and sub_val not in main[sub_key]:
-            return False
-    return True
-
-
-@typechecked
 def can_proceed(
     device: AutomatorDevice,
     expected_screennames: List[int],
@@ -305,12 +167,20 @@ def can_proceed(
 
     # end_screens = get end_screens()
     if not is_expected:
-        pprint(unpacked_screen_dict)
-        # TODO: Export the actual screen, screen data and expected screens in
+        # Export the actual screen, screen data and expected screens in
         # specific error log folder.
+        export_screen_data(
+            device=device,
+            screen_dict=unpacked_screen_dict,
+            script_description=script.script_description,
+            overwrite=True,
+            subdir="error",
+        )
         raise ReferenceError(
-            f"Error, the expected screen was not found in:{screen_nr}."
-            f"Searched for:{expected_screennames}."
+            f"Error, the expected screen was not found in:{screen_nr}. "
+            + f"Searched for:{expected_screennames}. The accompanying screen "
+            + "and xml can be found in:src/apkcontroller/<package_name>/<app_"
+            + f'version>/error/{script.script_description["screen_nr"]}.json'
         )
     return is_expected, screen_nr
